Add files via upload
Re-upload
Forgot to add stall input, usually if ran it in shell it would just terminate the program right after the calculations complete.
Added failsafe, pings MAL to check if user is correct, if not it resets the loop. (added error message for incorrect User, rather than it just crashing).